+++ src/main/java/com/gxdxx/instagram/service/token/AccessTokenCreateService.java
@@ -1,4 +1,4 @@
-package com.gxdxx.instagram.service;
+package com.gxdxx.instagram.service.token;
 
 import com.gxdxx.instagram.config.jwt.TokenProvider;
 import com.gxdxx.instagram.dto.response.SuccessResponse;
@@ -14,14 +14,14 @@ import org.springframework.transaction.annotation.Transactional;
 @RequiredArgsConstructor
 @Transactional
 @Service
-public class TokenService {
+public class AccessTokenCreateService {
 
     private final TokenProvider tokenProvider;
     private final RefreshTokenService refreshTokenService;
     private final UserRepository userRepository;
 
     // 전달받은 RefeshToken으로 토큰 유효성 검사를 진행하고, 유효한 토큰일 때 Refresh Token으로 사용자 id를 찾음
-    public SuccessResponse createNewAccessToken(String refreshToken, HttpServletResponse response) {
+    public SuccessResponse createAccessToken(String refreshToken, HttpServletResponse response) {
 
         // 토큰 유효성 검사에 실패하면 예외 발생
         if (!tokenProvider.validateToken(refreshToken)) {
